Add server registration worker
